Updated header
This is the code for the new header that we are planning on having. There is still some work to do, such as maybe add a search bar on the far right, and I know I gotta tidy up the code a little bit, but for now, this works. The color scheme might need some work, so we may need to revise that.