Updated with Footer
This is the updated header and footer design that we are planning on using for the site. This is just the basic design just so we can at least have a header and a footer. This is not mobile-responsive so we will have to make that mobile responsive, as well as revise the color scheme of course. And, Oliver also brought up a mis-alignment in the footer columns, so they will have to fix that as well(though it may not be necessary).